pyproject.toml
[tool.poetry]
name = "guess-the-episode"
version = "0.1.0"
description = "An online game where you guess the episode of a given frame from a TV show."
readme = "README.md"
authors = ["WolfDWyc <yoavwolfdw@gmail.com>",]
license = "AGPL-3.0-or-later"
classifiers = [
"License :: OSI Approved :: GNU Affero General Public License v3",
"Programming Language :: Python",
"Programming Language :: Python :: 3",
"Programming Language :: Python :: 3.11",
"Operating System :: OS Independent",
"Development Status :: 3 - Alpha"
]
keywords = ["chess"]

[tool.poetry.dependencies]
python = ">=3.11,<4.0"
pywebio = "==1.7.1"
pywebio-battery = "==0.4.0"
pydantic = "^2.3.0"
pillow = "^10.0.1"
annotated-types = "^0.5.0"
opencv-python = "^4.8.0.76"
tqdm = "^4.66.1"
decord = "^0.6.0"
fastapi = "^0.103.1"
uvicorn = "^0.23.2"
websockets = "^11.0.3"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.urls]
homepage = "https://github.com/WolfDWyc/ChessMaker"